[File] Move file.py outside pycore
