+++ nixio/pycore/__init__.py
@@ -8,4 +8,3 @@ from .tag import Tag
 from .multi_tag import MultiTag
 from .group import Group
 from .block import Block
-from .file import File, FileMode
